Give advice based on Factorio version
We now check the Factorio website for the current version, compare it to
the version running on our server, and change the advice to the user as
appropriate.

If the version of Factorio running is the same as the stable version, we
just tell people to install Factorio.

If the version is the current experimental version, we tell people to
opt into the experimental beta version of Factorio.

If it's an old version of Factorio, we tell users to download the
specific version.