Change non-zk colour from blue to green
Blue was too close to the real GOV.UK website, and very close to the
existing DWP corporate teal. This commit changes it to the other
complementary colour of the Zuzakistan colour, which is a slightly
off-putting green colour. Defra has green as their corporate colour, but
it's a proper green and not this sickly pea colour.